活动模块

+++ src/main/java/com/jgw/supercodeplatform/marketing/dto/activity/MarketingPrizeTypeParam.java
@@ -6,19 +6,18 @@ package com.jgw.supercodeplatform.marketing.dto.activity;
  */
 public class MarketingPrizeTypeParam {
 
-    private int id;//序列Id
+    private Long id;//序列Id
     private Integer prizeAmount;//金额数量,类型跟微信接口保持一致
     private Integer prizeProbability;//中奖几率
     private String prizeTypeName;//奖品类型名称
     private Byte randomAmount;//是否随机金额 1随机 0固定
-    
-	public int getId() {
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-	
 	public Integer getPrizeAmount() {
 		return prizeAmount;
 	}
